requirements.txt
delegator.py
celery
